strava/requirements.txt
#For more info on pip requirements see https://note.nkmk.me/en/python-pip-install-requirements/

icecream
requests
paho-mqtt